set taskHandle[cID] differently for each channel

+++ code/app.py
@@ -417,13 +417,13 @@ class RemApplication(QMainWindow):
             xLoc = (graphID * 300) + 50
             for chanID in range(2):
                 yLoc = (chanID * 200) + 120
-                self.listOfGraphs[chanID].append(DynamicGraphCanvas(self, width=3*self.scale, height=2*self.scale, dpi=100))
+                self.listOfGraphs[chanID].append(DynamicGraphCanvas(self, width=int(3 * self.scale), height=int(2 * self.scale), dpi=100))
                 self.listOfGraphs[chanID][graphID].move(int(xLoc * self.scale), int(yLoc * self.scale))
 
         self.setWindowTitle('Sleep stage classifier')
         xSize = self.graphNum * 310
         ySize = 550
-        self.resize(xSize * self.scale, ySize * self.scale)
+        self.resize(int(xSize * self.scale), int(ySize * self.scale))
         self.show()
         self.activateWindow()
         statusbar = self.statusBar()
